pyproject.toml
[project]
name = "latacc-medevac"
version = "0.1.0"
description = "LATACC MEDEVAC multi-agent system for NATO medical operations"
requires-python = ">=3.12"
authors = [{name = "Martinez Agullo, Pablo", email = "pablo.martinez.agullo@gmail.com"}]
license = "GPL-3.0-or-later"
readme = "README.md"

dependencies = [
    "cmop-observer",
    "latacc-common",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
cmop-observer = { workspace = true }
latacc-common = { workspace = true }
